Avoid constants when calculating max link length
The old version of this code took each candidate value and modified it
by applying arithmetic against constant values to find a maximum. The
operations being applied (addition and multiplication) are linear
transformations though, and therefore the order in which the logic takes
place can be turned around, finding the maxim candidate value first,
then transforming it.

This commit performs this change in order of operations. It first finds
the maximum value from the list of candidates without modifying them
(which lends itself better to instruction-level parallelisation), then
modifies the maximum value found. Special casing the fact that no result
could result from the reduce() call, in which case 0 is returned as in
the original implementation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>