Filter indices on iteration to avoid vector shuffling
Since the vector is not used afterwards, and this iteration is its only
usage after filtering, usage, there is no need to actually remove items
from its storage, only to filter them out as they are processed. This
has the benefit that it saves the shuffling of elements that occur
during retain().

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>