Use AoS for better cache usage
While the inputs to this function are given as separate arrays of
matching lengths, accessing their individual items from different
threads is not cache friendly, as each access on a different array means
that multiple cache lines end up being used, with much of their space
wasted.

This commit creates an array of structures (AoS) containing the same
input data, but transposed. This means that now the galaxies' data
needed by each thread (the galaxy being processed in that thread, and
then all galaxies it compares against) is contained sequentially in
memory, making better use of memory cache and creating less contention.
This of course comes at the cost of higher memory usage, but since the
normal input sizes are in the single million digits this isn't a big
cost to pay at all.

Local times look like this:

   # count        old        new    speedup
     52124   0.138703   0.090704  34.605596
    234118   1.462891   0.710315  51.444434
    536444   4.902282   2.460710  49.804805
    878420  10.608239   5.564962  47.541133
   1234186  18.205350   9.978118  45.191287
   1531862  26.878980  15.251497  43.258647
   1750764  38.158730  21.494584  43.670599
   1890144  45.333656  28.190827  37.814795
   1968522  56.581294  36.617912  35.282654
   2013862  72.646283  45.766789  37.000508

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>